+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonActivityLinkDTO.java
@@ -24,6 +24,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import pnet.data.api.util.AbstractLinkDTO;
+import pnet.data.api.util.ApprovalState;
 import pnet.data.api.util.WithBrandMatchcode;
 import pnet.data.api.util.WithCompanyId;
 import pnet.data.api.util.WithMatchcode;
@@ -55,6 +56,9 @@ public class PersonActivityLinkDTO extends AbstractLinkDTO
     @ApiModelProperty(notes = "True, if the employment has been approved already, false otherwise.")
     protected final boolean approved;
 
+    @ApiModelProperty(notes = "The current state of the audit process.")
+    protected final ApprovalState approvalState;
+
     @ApiModelProperty(notes = "The date and time from when this person has/had an actvitiy. "
         + "See https://github.com/porscheinformatik/pnet-data-api#validfromvalidto for additional information.")
     private final LocalDateTime validFrom;
@@ -70,8 +74,9 @@ public class PersonActivityLinkDTO extends AbstractLinkDTO
     public PersonActivityLinkDTO(@JsonProperty("tenant") String tenant, @JsonProperty("matchcode") String matchcode,
         @JsonProperty("companyId") Integer companyId, @JsonProperty("companyMatchcode") String companyMatchcode,
         @JsonProperty("companyNumber") String companyNumber, @JsonProperty("brandMatchcode") String brandMatchcode,
-        @JsonProperty("approved") boolean approved, @JsonProperty("validFrom") LocalDateTime validFrom,
-        @JsonProperty("validTo") LocalDateTime validTo, @JsonProperty("dueToFunction") boolean dueToFunction)
+        @JsonProperty("approved") boolean approved, @JsonProperty("approvalState") ApprovalState approvalState,
+        @JsonProperty("validFrom") LocalDateTime validFrom, @JsonProperty("validTo") LocalDateTime validTo,
+        @JsonProperty("dueToFunction") boolean dueToFunction)
     {
         super(tenant, matchcode);
 
@@ -80,6 +85,7 @@ public class PersonActivityLinkDTO extends AbstractLinkDTO
         this.companyNumber = companyNumber;
         this.brandMatchcode = brandMatchcode;
         this.approved = approved;
+        this.approvalState = approvalState;
         this.validFrom = validFrom;
         this.validTo = validTo;
         this.dueToFunction = dueToFunction;
@@ -126,6 +132,11 @@ public class PersonActivityLinkDTO extends AbstractLinkDTO
     public boolean isApproved()
     {
         return approved;
+    }
+
+    public ApprovalState getApprovalState()
+    {
+        return approvalState;
     }
 
     public LocalDateTime getValidFrom()
@@ -177,11 +188,10 @@ public class PersonActivityLinkDTO extends AbstractLinkDTO
     public String toString()
     {
         return String
-            .format(
-                "PersonActivityLinkDTO [companyId=%s, companyMatchcode=%s, companyNumber=%s, brandMatchcode=%s, "
-                    + "approved=%s, validFrom=%s, validTo=%s, dueToFunction=%s, tenant=%s, matchcode=%s]",
-                companyId, companyMatchcode, companyNumber, brandMatchcode, approved, validFrom, validTo, dueToFunction,
-                tenant, matchcode);
+            .format("PersonActivityLinkDTO [companyId=%s, companyMatchcode=%s, companyNumber=%s, brandMatchcode=%s, "
+                + "approved=%s, approvalState=%s, validFrom=%s, validTo=%s, dueToFunction=%s, tenant=%s, matchcode=%s]",
+                companyId, companyMatchcode, companyNumber, brandMatchcode, approved, approvalState, validFrom, validTo,
+                dueToFunction, tenant, matchcode);
     }
 
 }