+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonDataDTO.java
@@ -26,6 +26,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
+import pnet.data.api.util.ApprovalState;
 import pnet.data.api.util.WithId;
 import pnet.data.api.util.WithLastUpdate;
 import pnet.data.api.util.WithPersonId;
@@ -81,6 +82,9 @@ public class PersonDataDTO implements WithId, WithPersonId, WithTenants, WithLas
             + "ongoing (needed scope: SC_APPROVAL_PROCESS). This property is never null. If the scope is missing, "
             + "only approved persons will be available.")
     private boolean approved;
+
+    @ApiModelProperty(notes = "The current state of the audit process.")
+    private ApprovalState approvalState;
 
     @ApiModelProperty(notes = "The birthdate of the person (needed scope: SC_BIRTHDATE).")
     private LocalDate birthdate;
@@ -297,6 +301,16 @@ public class PersonDataDTO implements WithId, WithPersonId, WithTenants, WithLas
     public void setApproved(boolean approved)
     {
         this.approved = approved;
+    }
+
+    public ApprovalState getApprovalState()
+    {
+        return approvalState;
+    }
+
+    public void setApprovalState(ApprovalState approvalState)
+    {
+        this.approvalState = approvalState;
     }
 
     public LocalDate getBirthdate()
@@ -612,14 +626,15 @@ public class PersonDataDTO implements WithId, WithPersonId, WithTenants, WithLas
             .format(
                 "PersonDataDTO [personId=%s, administrativeTenant=%s, tenants=%s, formOfAddress=%s, "
                     + "academicTitle=%s, academicTitlePostNominal=%s, firstName=%s, lastName=%s, username=%s, "
-                    + "credentialsAvailable=%s, approved=%s, birthdate=%s, externalId=%s, guid=%s, preferredUserId=%s, "
-                    + "phoneNumber=%s, mobileNumber=%s, faxNumber=%s, email=%s, contactCompanyId=%s, costCenter=%s, "
-                    + "personnelNumber=%s, supervisorPersonnelNumber=%s, controllingArea=%s, personnelDepartment=%s, "
-                    + "jobDescription=%s, languages=%s, companies=%s, numbers=%s, functions=%s, activities=%s, "
-                    + "advisorAssignments=%s, portraitAvailable=%s, automaticDeletion=%s, checksum=%s, lastUpdate=%s]",
+                    + "credentialsAvailable=%s, approved=%s, approvalState=%s, birthdate=%s, externalId=%s, guid=%s, "
+                    + "preferredUserId=%s, phoneNumber=%s, mobileNumber=%s, faxNumber=%s, email=%s, "
+                    + "contactCompanyId=%s, costCenter=%s, personnelNumber=%s, supervisorPersonnelNumber=%s, "
+                    + "controllingArea=%s, personnelDepartment=%s, jobDescription=%s, languages=%s, companies=%s, "
+                    + "numbers=%s, functions=%s, activities=%s, advisorAssignments=%s, portraitAvailable=%s, "
+                    + "automaticDeletion=%s, checksum=%s, lastUpdate=%s]",
                 personId, administrativeTenant, tenants, formOfAddress, academicTitle, academicTitlePostNominal,
-                firstName, lastName, username, credentialsAvailable, approved, birthdate, externalId, guid,
-                preferredUserId, phoneNumber, mobileNumber, faxNumber, email, contactCompanyId, costCenter,
+                firstName, lastName, username, credentialsAvailable, approved, approvalState, birthdate, externalId,
+                guid, preferredUserId, phoneNumber, mobileNumber, faxNumber, email, contactCompanyId, costCenter,
                 personnelNumber, supervisorPersonnelNumber, controllingArea, personnelDepartment, jobDescription,
                 languages, companies, numbers, functions, activities, advisorAssignments, portraitAvailable,
                 automaticDeletion, checksum, lastUpdate);